Split GLM serving from rollout collection
